+++ src/components/FhiHeader/FhiHeader.stories.tsx
@@ -2,11 +2,8 @@
 /* eslint-disable react/jsx-props-no-spreading */
 
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import { faBars, faTimes } from '@fortawesome/free-solid-svg-icons';
 import FhiHeader from './FhiHeader';
 
-import fhiLogo from './assets/logo-norsk-kun-forkortelse-svart-uten-luft-rundt.svg';
-import fhiLogoLight from './assets/logo-norsk-kun-forkortelse-hvit-uten-luft-rundt.svg';
 
 export default {
   title: 'Fhi/FhiHeader',
@@ -18,16 +15,15 @@ const Template: ComponentStory<typeof FhiHeader> = (args) => <FhiHeader {...args
 export const Normal = Template.bind({});
 Normal.args = {
   projectName: 'Fhi React components',
-  logo: fhiLogo,
-  logoLight: fhiLogoLight,
   menuItems: [
     { name: 'Menu item 1', routerLink: '/1' },
     { name: 'Menu item 2', routerLink: '/2' },
     { name: 'Menu item 3', routerLink: '/3' },
     { name: 'Menu item 4', routerLink: '/4' },
   ],
-  faIcons: {
-    menu: faBars,
-    close: faTimes,
-  },
+};
+
+export const UtenMeny = Template.bind({});
+UtenMeny.args = {
+  projectName: 'Fhi React components',
 };