+++ src/components/FhiHeader/FhiHeader.tsx
@@ -1,7 +1,5 @@
 import { FC } from 'react';
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 
 import FhiMainMenu from './FhiMainMenu';
 
@@ -12,47 +10,29 @@ type FhiHeaderProps = {
     name: string,
     routerLink: string,
   }[],
-  logo: string,
-  logoLight: string,
-  faIcons: {
-    menu: IconDefinition,
-    close: IconDefinition,
-  }
 };
-
-const PaddingHeader = styled.header`
-background-color: ${({ theme, backgroundColor }) => backgroundColor || theme.fhiWhite};
-  margin-bottom: 0;
-  padding-bottom: 2.5rem;
-`;
 
 const FhiHeader:FC<FhiHeaderProps> = ({
   projectName,
   menuItems,
   className,
-  logo,
-  logoLight,
-  faIcons,
 }) => (
-  <PaddingHeader className={`fhi-app__header ${className}`}>
-    <div className="fhi-app__brand">
-      <div className="container fhi-app__brand-container">
-        <div className="fhi-app__brand-content">
-          <Link className="fhi-app__logo" to="/">
-            <img src={logo} alt="FHI-logo" />
+  <header className={`fhi-header ${className}`}>
+    <div className="fhi-header__brand">
+      <div className="container fhi-header__brand-container">
+        <div className="fhi-header__brand-content">
+          <Link to="/" className="fhi-header__logo">
+            <i className="icon-fhi-logo fhi-header__logo-icon" />
+            <span className="visually-hidden">{projectName}</span>
           </Link>
-          <div className="fhi-app__project">
-            <span className="fhi-app__project-name">{projectName}</span>
+          <div className="fhi-header__project">
+            <span className="fhi-header__project-name">{projectName}</span>
           </div>
         </div>
       </div>
     </div>
-    <FhiMainMenu
-      menuItems={menuItems}
-      logo={logoLight}
-      faIcons={faIcons}
-    />
-  </PaddingHeader>
+    <FhiMainMenu menuItems={menuItems} />
+  </header>
 );
 
 FhiHeader.defaultProps = {